backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
Flask-Cors==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1